- mdmkolbe fixed several warnings including:     - Copy constructors not calling constructors of virtual parrents,     - String formats not matching argument types,     - Missing size_t cast before a -1, and     - Missing #if's around uncalled functions on some configurations.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@23203 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/win32/win32.cpp
@@ -455,7 +455,7 @@ Win32Assistant::Win32Assistant (iObjectRegistry* r)
     
     windowClass = new uint8[(wndClass.Length()+1) * sizeof (WCHAR)];
     csUnicodeTransform::UTF8toWC ((wchar_t*)windowClass,
-      wndClass.Length()+1, (utf8_char*)(wndClass.GetData()), -1);
+      wndClass.Length()+1, (utf8_char*)(wndClass.GetData()), (size_t)-1);
 
     wc.cbSize	      = sizeof (wc);
     wc.hCursor        = 0;
